Save JSON as CouchDB documents directly
This allows to run map/reduce on them for example. This also fixes the
issue with UTF-8 documents, without having to force the encoding

The API test suite currently has two failing tests because I have to
rewrite the CouchDB JSON to remove the built-in keys (starting with
underscores, like _rev and _id) that CouchDB adds to documents. The
rewritten files don't have spaces between the keys and values because of
the implementation of the JSON module